+++ GloomyDiary/Model/Talking.swift
